refactor(*): use zeitwerk for autoloading
